refactor: :recycle: i18n resource src/shared/constant폴더로 이동
