+++ src/modules/i18n/index.ts
@@ -1,24 +1,7 @@
-import i18n, { Resource } from 'i18next';
+import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 
-const resources: Resource = {
-  en: {
-    translation: {
-      hello: 'Hello',
-      helloUser: 'Hello {name}',
-      helloTag: '<h1>Hello</h1>',
-      helloLineBreak: 'hello\nhi',
-    },
-  },
-  ko: {
-    translation: {
-      hello: '안녕',
-      helloUser: '안녕 {{name}}',
-      helloTag: '<h1>안녕</h1>',
-      helloLineBreak: '안녕\n안뇽',
-    },
-  },
-};
+import resources from '@src/shared/constant/i18nResource';
 
 i18n.use(initReactI18next).init({
   resources,
